hash: breaking api change
the hash interfaces returned errors before, but the underlying
implementations picked for the hash package never return errors on calls
to Write.

bubbling the errors in the interface like this leads to a lot of
unnecessary error handling, so it's worth breaking the api contract here
(and we're pre-1.0.0 anyway).